refactor(device-api): split flagcx_kernel.h into core and internal headers
